Remove non factory functions except for startBattle
In order to follow the exercise instructions, I am keeping everything
that is not the startBattle function nor the factory functions within
the startBattle function.

Also, have the grids built for both players and adding click events to
player 2's spaces. It is working but need to get the while loop set up
so the game will go back and forth between the human and AI players.

[Ticket: X]