Prevent repeatedly attacked spots by computer
The clickability/click event for human players prevents them from
selecting a position that was already picked. This was not the case for
computer players. The attack function is now recursive if the randomly
selected attack by the computer returns an invalid result, so that
a valid spot eventually gets attacked successfully.